requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-genai==0.3.0
Pillow==10.1.0
PyMuPDF==1.23.14
python-docx==1.1.0
pdf2image==1.17.0
Werkzeug==2.3.7
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
